% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_function.R
\name{MCS}
\alias{MCS}
\title{Monte Carlo Simulation
This function recursively generates the values of factorials.}
\usage{
MCS(initial, mean.logret, sd.logret, steps, paths)
}
\arguments{
\item{k}{a scalar determining the factorials.}
}
\value{
a scalar equivalent to \eqn{n!}.
}
\description{
Monte Carlo Simulation
This function recursively generates the values of factorials.
}
\examples{
## a trivial example
frac(10)
prod(1:10)
}
\references{

}
\seealso{
\code{prod}
}
